Create mixin for actor sheets
