+++ module/applications/actor/base-sheet.mjs
@@ -20,12 +20,13 @@ import TraitSelector from "./trait-selector.mjs";
 import ProficiencyConfig from "./proficiency-config.mjs";
 import ToolSelector from "./tool-selector.mjs";
 import { simplifyBonus } from "../../utils.mjs";
+import { ActorSheetMixin } from "./sheet-mixin.mjs";
 
 /**
  * Extend the basic ActorSheet class to suppose system-specific logic and functionality.
  * @abstract
  */
-export default class ActorSheet5e extends ActorSheet {
+export default class ActorSheet5e extends ActorSheetMixin(ActorSheet) {
 
   /**
    * Track the set of item filters which are applied
@@ -757,97 +758,6 @@ export default class ActorSheet5e extends ActorSheet {
   /* -------------------------------------------- */
 
   /**
-   * Prepare an array of context menu options which are available for owned ActiveEffect documents.
-   * @param {ActiveEffect5e} effect         The ActiveEffect for which the context menu is activated
-   * @returns {ContextMenuEntry[]}          An array of context menu options offered for the ActiveEffect
-   * @protected
-   */
-  _getActiveEffectContextOptions(effect) {
-    return [
-      {
-        name: "DND5E.ContextMenuActionEdit",
-        icon: "<i class='fas fa-edit fa-fw'></i>",
-        callback: () => effect.sheet.render(true)
-      },
-      {
-        name: "DND5E.ContextMenuActionDuplicate",
-        icon: "<i class='fas fa-copy fa-fw'></i>",
-        callback: () => effect.clone({label: game.i18n.format("DOCUMENT.CopyOf", {name: effect.label})}, {save: true})
-      },
-      {
-        name: "DND5E.ContextMenuActionDelete",
-        icon: "<i class='fas fa-trash fa-fw'></i>",
-        callback: () => effect.deleteDialog()
-      },
-      {
-        name: effect.disabled ? "DND5E.ContextMenuActionEnable" : "DND5E.ContextMenuActionDisable",
-        icon: effect.disabled ? "<i class='fas fa-check fa-fw'></i>" : "<i class='fas fa-times fa-fw'></i>",
-        callback: () => effect.update({disabled: !effect.disabled})
-      }
-    ];
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Prepare an array of context menu options which are available for owned Item documents.
-   * @param {Item5e} item                   The Item for which the context menu is activated
-   * @returns {ContextMenuEntry[]}          An array of context menu options offered for the Item
-   * @protected
-   */
-  _getItemContextOptions(item) {
-
-    // Standard Options
-    const options = [
-      {
-        name: "DND5E.ContextMenuActionEdit",
-        icon: "<i class='fas fa-edit fa-fw'></i>",
-        callback: () => item.sheet.render(true)
-      },
-      {
-        name: "DND5E.ContextMenuActionDuplicate",
-        icon: "<i class='fas fa-copy fa-fw'></i>",
-        condition: () => !["race", "background", "class", "subclass"].includes(item.type),
-        callback: () => item.clone({name: game.i18n.format("DOCUMENT.CopyOf", {name: item.name})}, {save: true})
-      },
-      {
-        name: "DND5E.ContextMenuActionDelete",
-        icon: "<i class='fas fa-trash fa-fw'></i>",
-        callback: () => item.deleteDialog()
-      }
-    ];
-
-    // Toggle Attunement State
-    if ( ("attunement" in item.system) && (item.system.attunement !== CONFIG.DND5E.attunementTypes.NONE) ) {
-      const isAttuned = item.system.attunement === CONFIG.DND5E.attunementTypes.ATTUNED;
-      options.push({
-        name: isAttuned ? "DND5E.ContextMenuActionUnattune" : "DND5E.ContextMenuActionAttune",
-        icon: "<i class='fas fa-sun fa-fw'></i>",
-        callback: () => item.update({
-          "system.attunement": CONFIG.DND5E.attunementTypes[isAttuned ? "REQUIRED" : "ATTUNED"]
-        })
-      });
-    }
-
-    // Toggle Equipped State
-    if ( "equipped" in item.system ) options.push({
-      name: item.system.equipped ? "DND5E.ContextMenuActionUnequip" : "DND5E.ContextMenuActionEquip",
-      icon: "<i class='fas fa-shield-alt fa-fw'></i>",
-      callback: () => item.update({"system.equipped": !item.system.equipped})
-    });
-
-    // Toggle Prepared State
-    if ( ("preparation" in item.system) && (item.system.preparation?.mode === "prepared") ) options.push({
-      name: item.system?.preparation?.prepared ? "DND5E.ContextMenuActionUnprepare" : "DND5E.ContextMenuActionPrepare",
-      icon: "<i class='fas fa-sun fa-fw'></i>",
-      callback: () => item.update({"system.preparation.prepared": !item.system.preparation?.prepared})
-    });
-    return options;
-  }
-
-  /* -------------------------------------------- */
-
-  /**
    * Initialize Item list filters by activating the set of filters which are currently applied
    * @param {number} i  Index of the filter in the list.
    * @param {HTML} ul   HTML object for the list item surrounding the filter.
@@ -863,24 +773,6 @@ export default class ActorSheet5e extends ActorSheet {
 
   /* -------------------------------------------- */
   /*  Event Listeners and Handlers                */
-  /* -------------------------------------------- */
-
-  /**
-   * Handle input changes to numeric form fields, allowing them to accept delta-typed inputs.
-   * @param {Event} event  Triggering event.
-   * @protected
-   */
-  _onChangeInputDelta(event) {
-    const input = event.target;
-    const value = input.value;
-    if ( ["+", "-"].includes(value[0]) ) {
-      const delta = parseFloat(value);
-      const item = this.actor.items.get(input.closest("[data-item-id]")?.dataset.itemId);
-      if ( item ) input.value = Number(foundry.utils.getProperty(item, input.dataset.name)) + delta;
-      else input.value = Number(foundry.utils.getProperty(this.actor, input.name)) + delta;
-    } else if ( value[0] === "=" ) input.value = value.slice(1);
-  }
-
   /* -------------------------------------------- */
 
   /**
@@ -1133,26 +1025,6 @@ export default class ActorSheet5e extends ActorSheet {
   /* -------------------------------------------- */
 
   /**
-   * Stack identical consumables when a new one is dropped rather than creating a duplicate item.
-   * @param {object} itemData         The item data requested for creation.
-   * @returns {Promise<Item5e>|null}  If a duplicate was found, returns the adjusted item stack.
-   */
-  _onDropStackConsumables(itemData) {
-    const droppedSourceId = itemData.flags.core?.sourceId;
-    if ( itemData.type !== "consumable" || !droppedSourceId ) return null;
-    const similarItem = this.actor.items.find(i => {
-      const sourceId = i.getFlag("core", "sourceId");
-      return sourceId && (sourceId === droppedSourceId) && (i.type === "consumable") && (i.name === itemData.name);
-    });
-    if ( !similarItem ) return null;
-    return similarItem.update({
-      "system.quantity": similarItem.system.quantity + Math.max(itemData.system.quantity, 1)
-    });
-  }
-
-  /* -------------------------------------------- */
-
-  /**
    * Adjust the preparation mode of a dropped spell depending on the drop location on the sheet.
    * @param {object} itemData    The item data requested for creation. **Will be mutated.**
    */
@@ -1249,39 +1121,7 @@ export default class ActorSheet5e extends ActorSheet {
     parent.appendChild(input);
   }
 
-  /* -------------------------------------------- */
 
-  /**
-   * Change the uses amount of an Owned Item within the Actor.
-   * @param {Event} event        The triggering click event.
-   * @returns {Promise<Item5e>}  Updated item.
-   * @protected
-   */
-  async _onUsesChange(event) {
-    event.preventDefault();
-    const itemId = event.currentTarget.closest(".item").dataset.itemId;
-    const item = this.actor.items.get(itemId);
-    const uses = Math.clamped(0, parseInt(event.target.value), item.system.uses.max);
-    event.target.value = uses;
-    return item.update({"system.uses.value": uses});
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Change the quantity of an Owned Item within the actor.
-   * @param {Event} event        The triggering click event.
-   * @returns {Promise<Item5e>}  Updated item.
-   * @protected
-   */
-  async _onQuantityChange(event) {
-    event.preventDefault();
-    const itemId = event.currentTarget.closest(".item").dataset.itemId;
-    const item = this.actor.items.get(itemId);
-    const quantity = Math.max(0, parseInt(event.target.value));
-    event.target.value = quantity;
-    return item.update({"system.quantity": quantity});
-  }
 
   /* -------------------------------------------- */
 
@@ -1311,33 +1151,6 @@ export default class ActorSheet5e extends ActorSheet {
     const itemId = event.currentTarget.closest(".item").dataset.itemId;
     const item = this.actor.items.get(itemId);
     return item.rollRecharge();
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Handle toggling and items expanded description.
-   * @param {Event} event   Triggering event.
-   * @private
-   */
-  async _onItemSummary(event) {
-    event.preventDefault();
-    const li = $(event.currentTarget).parents(".item");
-    const item = this.actor.items.get(li.data("item-id"));
-    const chatData = await item.getChatData({secrets: this.actor.isOwner});
-
-    // Toggle summary
-    if ( li.hasClass("expanded") ) {
-      const summary = li.children(".item-summary");
-      summary.slideUp(200, () => summary.remove());
-      this._expanded.delete(item.id);
-    } else {
-      const summary = $(await renderTemplate("systems/dnd5e/templates/items/parts/item-summary.hbs", chatData));
-      li.append(summary.hide());
-      summary.slideDown(200);
-      this._expanded.add(item.id);
-    }
-    li.toggleClass("expanded");
   }
 
   /* -------------------------------------------- */
